requirements.txt
python-telegram-bot==9.0.0
pyTelegramBotAPI==3.5.2